Add example usage of time-dependent GEV fit
The example uses linear trends and annual cycles, and shows how the test
based on the deviance statistic can be used to test for significance.  A
later example will also include the NAO index.

The time-independent example is adapted to be more similar to the new
script.  The README file is updated as well, to include both examples
and to give more information on how to run them.