Add example usage of GEV fit with NAO-dependence
The new example script is similar to Time-dependent_GEV_fit.py, which
has been modified slightly, and shows in addition how to include a
linear dependence on a climate index like NAO in the GEV model.  As
usual, the data files for the climate index need to be downloaded
separately, as explained in tools_climate.py.