Make high tide surge the default for load_data
By making high tide skew surge levels the default in the load_data
function, the code became cleaner, shorter, and easier to understand.
This has almost no effect for the GESLA-2 dataset, since high tide surge
has always been the default for it, but it makes it easier to switch to
the surge dataset of Reinert et al. (2021), for which surge levels at
both high and low tide are available.  If low tide surge levels are not
explicitly requested, load_data returns now only high tide surge.  The
behaviour of the function is thus more similar for the two datasets.