Streamline the code
Three of the scripts that reproduce the main results of the manuscript
use now our surge dataset by default instead of the GESLA-2 dataset,
just like it is done in the manuscript.  This allowed to clean up the
code and to get rid of extra lines that were only required for the
GESLA-2 dataset.  For the two code examples that exemplify the general
usage of this code package, the GESLA-2 dataset is kept the default,
because it can be easily obtained by everyone from the internet.

Information on citing our paper is added to the README.