Implement a better confidence interval computation
This commit introduces a new way of computing the confidence interval
for a non-time-dependent GEV distribution.  The new functions follow
the book by Coles (2001).  The function that was previously used to
compute the confidence interval is removed, as it is not needed in any
of the sample scripts.  The function to print the GEV parameters is
moved from tools_errors to advanced_GEV_analysis.  The former file is
removed because it became empty.

The new way to calculate confidence intervals is mathematically more
sound and generally leads to smaller confidence intervals, thus it
improves the results.  The old way of computing confidence intervals
(now removed) gave a more conservative estimate.

Note that this change does not impact time-dependent GEV distributions
as discussed in the paper.

This modification has been proposed by Marvin Lorenz (IOW), who also
implemented a first version of the new methods.

A new sample script that uses annual maxima instead of monthly maxima is
added with this commit.  Since no conversion between return periods in
years and months is necessary for annual maxima, this code is slightly
cleaner and makes the principles of this new method clearer.